Move `Sequence`'s `removingDuplicates` helper to `Array+Helpers` with unit tests.
